+++ org-roam.el
@@ -167,18 +167,19 @@ If FILE is not specified, use the current buffer's file-path."
   (if-let ((path (or file
                   (buffer-file-name))))
       (save-match-data
-	(org-roam--org-file-p path)
-	(f-descendant-of-p (file-truename path)
-			   (file-truename org-roam-directory)))))
+        (and
+         (org-roam--org-file-p path)
+         (f-descendant-of-p (file-truename path)
+                            (file-truename org-roam-directory))))))
 
 (defun org-roam--list-files (dir)
   "Return all Org-roam files located within DIR, at any nesting level.
 Ignores hidden files and directories."
   (let ((regex (concat "\\.\\(?:"(mapconcat #'regexp-quote org-roam-file-extensions "\\|" )"\\)\\(?:\\.gpg\\)?\\'"))
-	result)
+        result)
     (dolist (file (directory-files-recursively dir regex) result)
       (when (and (file-readable-p file) (org-roam--org-file-p file))
-	(push file result)))))
+        (push file result)))))
 
 (defun org-roam--list-all-files ()
   "Return a list of all Org-roam files within `org-roam-directory'."